# Refactor(eos_designs): Optimize and fix switch facts