--- a/ansible_collections/arista/avd/roles/eos_designs/tasks/main.yml
+++ b/ansible_collections/arista/avd/roles/eos_designs/tasks/main.yml
@@ -35,13 +35,6 @@
   check_mode: false
   run_once: true
 
-- name: Set eos_designs facts per device
-  tags: [build, provision, facts]
-  ansible.builtin.set_fact:
-    switch: "{{ avd_switch_facts[inventory_hostname].switch }}"
-  delegate_to: localhost
-  changed_when: false
-
 - name: Generate YAML file with hostvars (only for debugging)
   tags: [debug, never]
   ansible.builtin.template:
@@ -98,7 +91,7 @@
 
 - name: Remove avd_switch_facts
   tags: [build, provision, facts, remove_avd_switch_facts]
-  ansible.builtin.set_fact:
+  arista.avd.set_vars:
     avd_switch_facts: null
   run_once: true
   check_mode: false