--- a/ansible_collections/arista/avd/tests/sanity/ignore-2.14.txt
+++ b/ansible_collections/arista/avd/tests/sanity/ignore-2.14.txt
@@ -2,6 +2,7 @@ plugins/modules/batch_template.py validate-modules:missing-gplv3-license
 plugins/modules/configlet_build_config.py validate-modules:missing-gplv3-license
 plugins/modules/eos_designs_facts.py validate-modules:missing-gplv3-license
 plugins/modules/inventory_to_container.py validate-modules:missing-gplv3-license
+plugins/modules/set_vars.py validate-modules:missing-gplv3-license
 plugins/modules/validate_and_template.py validate-modules:missing-gplv3-license
 plugins/modules/verify_requirements.py validate-modules:missing-gplv3-license
 plugins/modules/yaml_templates_to_facts.py validate-modules:missing-gplv3-license