## Change Summary

<!-- Enter short PR description -->
Optimize and fix switch facts

## Related Issue(s)

When running `eos_designs` the last task is the clear `avd_switch_facts`. The purpose is to remove this very big dataset from the variable space, speeding up all following tasks significantly.

The remove task was using `set_fact` to clear `avd_switch_facts`, which had been set from `eos_designs_facts` action plugin. 

Unfortunately `set_fact` only cleared the _var_ `avd_switch_facts` but _not_ the `ansible_facts.avd_switch_facts` fact. This is just how `set_fact` works, and not anything we can change.

Ansible does not provide any other standard modules to set a fact _and_ var similar to how modules normally do, so part of this PR is a new plugin `arista.avd.set_vars` which returns any arguments into `ansible_facts` which Ansible in turn sets as _vars_.

Since this is the same way `eos_designs_facts` sets `avd_switch_facts` in the first place, we can use this method to overwrite both copies will `none`.

## Component(s) name

`arista.avd.eos_designs`
`arista.avd.set_vars`
`arista.avd.yaml_templates_to_facts`

## Proposed changes
<!--- Describe your changes in detail -->
<!--- Describe data model implemented for new features -->

- New action plugin `arista.avd.set_vars` to offer a simple way to set facts with a module.
- Use `set_vars` instead of `set_fact` for clearing `avd_switch_facts`
- Move setting of `switch.*` from `avd_switch_facts[<hostname].switch` into `yaml_templates_to_facts`.
  Partially for speed and partially for future option to skip this.
- Clean up unused and undocumented `remove_avd_switch_facts` option from `yaml_templates_to_facts`

## How to test
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->

Running eos_designs_unit_tests. Ignore the `real` time, which was likely some thermal throttling, but `user` shows 25% faster.

```sh
#devel
real    5m6,659s
user    39m27,159s
sys     1m9,590s

#PR
real    3m8,249s
user    30m24,396s
sys     0m52,351s
```

## Checklist

### User Checklist

<!-- Add your own checklist using MD syntax and by replacing N/A -->
- N/A

### Repository Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My code has been rebased from devel before I start
- [ ] I have read the [**CONTRIBUTING**](https://avd.sh/en/latest/docs/contribution/overview.html) document.
- [ ] My change requires a change to the documentation and documentation have been updated accordingly.
- [ ] I have updated [molecule CI](https://github.com/aristanetworks/ansible-avd/tree/devel/ansible_collections/arista/avd/molecule) testing accordingly. (check the box if not applicable)